mail/postfix: Fix typos in filenames and variables for header_checks
Use "Headerchecks" or "headerchecks" everywhere.
