+++ mail/postfix/src/opnsense/mvc/app/models/OPNsense/Postfix/Menu/Menu.xml
@@ -8,7 +8,7 @@
       <Senders url="/ui/postfix/sender/index" order="40"/>
       <Senderbcc VisibleName="Sender BCC" url="/ui/postfix/senderbcc/index" order="45"/>
       <Sendercanonical VisibleName="Sender Canonical Rewriting" url="/ui/postfix/sendercanonical/index" order="48"/>
-      <HeaderChecks VisibleName="Header Checks" url="/ui/postfix/headerchecks/index" order="49"/>
+      <Headerchecks VisibleName="Header Checks" url="/ui/postfix/headerchecks/index" order="49"/>
       <Address VisibleName="Address Rewriting" url="/ui/postfix/address/index" order="50"/>
       <LogFile VisibleName="Log File" order="50" url="/ui/diagnostics/log/core/mail"/>
     </Postfix>
