+++ mail/postfix/src/opnsense/service/templates/OPNsense/Postfix/header_checks_delivering
@@ -1,6 +1,6 @@
 {% if helpers.exists('OPNsense.postfix.general.enabled') and OPNsense.postfix.general.enabled == '1' %}
-{%   if helpers.exists('OPNsense.postfix.headercheck.headerchecks.headercheck') %}
-{%     for headercheck_list in helpers.toList('OPNsense.postfix.headercheck.headerchecks.headercheck') %}
+{%   if helpers.exists('OPNsense.postfix.headerchecks.headerchecks.headercheck') %}
+{%     for headercheck_list in helpers.toList('OPNsense.postfix.headerchecks.headerchecks.headercheck') %}
 {%       if headercheck_list.enabled == '1' and headercheck_list.filter == 'WHILE_DELIVERING'%}
 {{ headercheck_list.expression }}
 {%       endif %}
